requirements.txt
langgraph
requests
python-dotenv
langchain
langchain-community
streamlit
langgraph-checkpoint-sqlite
langchain_groq
tavily-python
langchain_ollama
pdfplumber
langchain_cohere
langsmith
unstructured 
pandas 
bs4 
openpyxl 
python-pptx